Downgrade Selenium from 3.3.0 to 3.2.0
Looking at the hub and node logs with 3.3.0 it turned out that the Chrome node
was constantly registering and deregistering with the hub. This was causing the
acceptance setup to timeout when trying to initialize a new session because the
hub was waiting for the node to register back.

There are a couple of open issues that confirm this issue:

- https://github.com/SeleniumHQ/docker-selenium/issues/425 and
- https://github.com/SeleniumHQ/selenium/issues/3744.